Remove use of dependencies in trait definitions. Split crate.
I realized that because I link to fuzzcheck as a pre-compiled binary, 
the compiler copies fuzzcheck’s dependencies and consider them different
to any other dependency in the crate, even if they share the same name,
version, API, and implementation. As a consequence, if I declare traits 
in fuzzcheck with foreign types in them, such as `rand() -> ThreadRng`,
it is impossible for an other crate to implement them! So I removed all 
those instances where the API of fuzzcheck relied on a crate. 

I then took the opportunity to move InputGenerator and VectorGenerator
to their own crate, that will contain more generators. 

Note: at some point I hesitated getting rid of InputGenerator and 
working with the individual closure directly (for example, 
`Complexity: Fn(&T) -> f64`). It was very verbose but kind of beautiful 
and more granular. I scrapped it in the end but maybe I'll revert back
to that.