Fix compilation settings for Linux
The gold linker is used when it is available, and the
force-frame-pointer option is set to yes.

A new option is added to `cargo fuzzcheck init` to specify a custom
fuzzcheck repository to clone.

cargo-fuzzcheck has a new `clean` command that runs cargo clean for
the main library, the fuzzcheck library, and the fuzz targets