Huge rewrite with no history
I am very sorry about this.
So, basically, many things changed:

FuzzedInput was split into Mutator and Serializer.

Mutator’s interface has been completely reworked. Now I distinguish
between the value, cache, and mutation step. An explanation of these
is included in the documentation. There is also fewer annoying
requirements on it, for example now there is no need to specify how to
hash a value. Instead, the hash is created from its encoding.

I worked on some of the basic Mutators and added a tuple and either one,
as a proof of concept of what proper generic mutators could be. But all
included mutators are still very basic and lacking in quality.

The Serializer now supports serde, optionally, and is thought more as a
property of World. I also made a ByteSerializer. The extension of
artifacts is customizable.

It is sooooo much faster now! I rewrote the data structure of the Pool
such that it stores a sorted list of features and uses Slab a lot as
a kind of dense storage with O(1) lookup. Also, the Pool uses property
tests now, which I have found are much better at finding bugs than my
previous test suite. It is also easier to refactor it.

“Feature Groups” has been implemented, kind of. It is a new way to
score features that share a common characteristic, such as edge features
that have an identical PC but different counters. Now all these features
share a common score and split it among themselves, which I thought was
fairer. This will also enable a more fine-grained recording of code
coverage because now I don't have to fear that, say, distinguishing
between more than 8 counters of code edges would wrongly inflate their
scores.

The compilation steps have changed a bit too, to avoid conflicts in
common dependencies.

I reorganized the crate’s files so that its generated documentation
would look better, and I think it makes sense. Speaking of
documentation, it has improved a bit, but it is still not great to be
honest.