Add concept of mutator supertype
A mutator is a subtype of another if:
* anywhere the supertype is used, the subtype can be substituted
* the subtype can always be upcasted to the supertype

Concretely, I wanted this because we currently have
“single variant mutators” for generating only one variant of an enum.

So given the enum:
```rust
enum X {
    A(u8)
    B(u16, Vec<u8>)
    C {}
    D
}
```
We can make a:
```rust
enum SingleVariantXMutator<A, B, C, D> {
    A(A),
    B(B),
    C(C),
    D(D)
}
```
where A, B, C, D are the mutators used for each variant. So a
SingleVariantXMutator::A(..) will only ever generate values of the
form X::A(..).

To prove this property at compile time, we can create a helper method:
```rust
fn make_variant_a_mutator<A>(m: A) -> SingleVariantXMutator<A, !, !, !>
```
The Never type used in the type params communicate that the single
variant is incapable of producing values other than X::A

But now that we have all that, we'd like to compose different
single-variant mutators into many-variants mutators. Say we have a
SingleVariantMutator<A, !, !, !> and a SingleVariantMutator<!, B, !, !>,
and we want a SingleVariantMutator<A, B, !, !>. This is what the
concepts of supertype and common supertype help.

The rule is as follows: SingleVariant<A1, B1, C1, D1> is a supertype of
SingleVariant<A2, B2, C2, D2> if all the type params of its variant
mutators (A1, ..) are supertypes of their counterparts (A2, ..).

We also automatically implement some relations:
* Any mutator M is a supertype of NeverMutator (equivalent to !)
* A mutator is its own supertype

And finally, we add another trait for finding the strictest (preferably)
common supertype of two mutators.

For example, SingleVariant<A, B, !, !> is a common supertype of
SingleVariant<A, !, !, !> and SingleVariant<!, B, !, !>. We can find it
by writing
```rust
<
    SingleVariantX<A, !, !, !>
    as
    CommonMutatorSuperType<
        X,
        SingleVariantX<!, B, !, !>
    >
>::Output
```

In the end, I hope to be able to replace the current system for
generating enum mutators with a composition of single-variant mutators.
This would be good because:
* single-variant mutators are very easy to write
* it would enable creating enum mutators that do not generate
certain variants
* I can pull the enum mutator logic out of the proc macro and into a
more general AlternationMutator<..> (more on that later)