Get rid of grammar-based string mutators
It has a fatal bug, which I could probably easily fix. But I also plan
on changing a lot of other mutators, and every time I do so, I will
need to update the incremental map conformances that are needed to build
the grammar-based string mutator, so it's not worth it.

I also want the parser to give all of the possible interpretations of
the string to the mutator, which it doesn't do currently. That's more
work that I don't want to do now.

In the meantime, one can use a MapMutator from AST to String